Reworked ConcurrentRequestScheduler and fixed concurrency defect; cleaned up LoadTest a little

+++ og-configuration/src/main/java/com/cleversafe/og/guice/JsonModule.java
@@ -463,7 +463,7 @@ public class JsonModule extends AbstractModule
    {
       final ConcurrencyConfig concurrency = this.config.getConcurrency();
       if (ConcurrencyType.THREADS == concurrency.getType())
-         return new ConcurrentRequestScheduler(Math.round(concurrency.getCount()));
+         return new ConcurrentRequestScheduler((int) Math.round(concurrency.getCount()));
       else
       {
          final Distribution count = new UniformDistribution(concurrency.getCount(), 0.0);
